+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/cargo-trend/compare/v0.1.1...Unreleased) - ReleaseDate
 
+* [Fixed] remove unwrap
+
 ## [v0.1.1](https://github.com/dalance/cargo-trend/compare/v0.1.0...v0.1.1) - 2019-10-04
 
-* Fix subcommand parsing
+* [Fixed] Fix subcommand parsing